+++ packages/hydrogen/src/components/CartLinePrice/CartLinePrice.client.tsx
@@ -1,9 +1,9 @@
 import React, {ElementType} from 'react';
 import {useCartLine} from '../CartLineProvider';
-import {Money} from '../Money';
+import {Money, MoneyProps} from '../Money';
 import {Props} from '../types';
 
-export interface CartLinePriceProps {
+export interface CartLinePriceProps extends Omit<MoneyProps, 'money'> {
   /** The type of price. Valid values:`regular` (default) or `compareAt`. */
   priceType?: 'regular' | 'compareAt';
 }
@@ -34,8 +34,6 @@ export function CartLinePrice<TTag extends ElementType>(
         amount: price.amount * cartLine.quantity,
         currencyCode: price.currencyCode,
       }}
-    >
-      {props.children}
-    </Money>
+    />
   );
 }